backend/requirements.txt
fastapi
uvicorn
python-socketio
tensorflow-cpu==2.15.0
numpy